refactor: Move bases classes
Move base classes Node and LinkedList into a base package to share with other types of linked list implementations.
